histour-chatbot/requirements.txt
fastapi
uvicorn
openai==0.28
faiss-cpu
sentence-transformers
psycopg2-binary
python-dotenv
pandas